@@ -11,49 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.8",
-    "@angular/cdk": "^7.3.4",
-    "@angular/common": "^7.2.8",
-    "@angular/compiler": "^7.2.8",
-    "@angular/core": "^7.2.8",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.8",
-    "@angular/material": "^7.3.4",
-    "@angular/platform-browser": "^7.2.8",
-    "@angular/platform-browser-dynamic": "^7.2.8",
-    "@angular/router": "^7.2.8",
-    "@aspnet/signalr": "^1.1.2",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.15",
-    "@fortawesome/free-solid-svg-icons": "^5.7.2",
-    "core-js": "^2.6.5",
+    "@angular/animations": "^8.2.9",
+    "@angular/cdk": "^8.2.2",
+    "@angular/common": "^8.2.9",
+    "@angular/compiler": "^8.2.9",
+    "@angular/core": "^8.2.9",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.9",
+    "@angular/material": "^8.2.2",
+    "@angular/platform-browser": "^8.2.9",
+    "@angular/platform-browser-dynamic": "^8.2.9",
+    "@angular/router": "^8.2.9",
+    "@fortawesome/angular-fontawesome": "^0.5.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.25",
+    "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "@microsoft/signalr": "^3.0.0",
+    "core-js": "^3.2.1",
     "hammerjs": "^2.0.8",
     "moment": "^2.24.0",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.29"
+    "rxjs": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.5",
-    "@angular/cli": "^7.3.5",
-    "@angular/compiler-cli": "^7.2.8",
-    "@angular/language-service": "^7.2.8",
-    "@types/jasmine": "^2.8.16",
+    "@angular-devkit/build-angular": "~0.803.8",
+    "@angular/cli": "^8.3.8",
+    "@angular/compiler-cli": "^8.2.9",
+    "@angular/language-service": "^8.2.9",
+    "@types/jasmine": "^3.4.2",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^10.12.30",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "^12.7.11",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "prettier": "^1.16.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "prettier": "^1.18.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.1",
-    "tslint": "^5.13.1",
-    "typescript": "~3.1.6"
+    "ts-node": "~8.4.1",
+    "tslint": "^5.20.0",
+    "typescript": "~3.5.3"
   },
   "prettier": {
     "printWidth": 140,
