@@ -8,7 +8,7 @@ import {
   QueryList,
   ViewChildren,
 } from '@angular/core';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import { HubService } from 'src/app/services/hub.service';
 
 import { PlaceDirective } from '../../../directives/place.directive';
