@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import * as signalR from '@aspnet/signalr';
+import * as signalR from '@microsoft/signalr';
 import { BehaviorSubject } from 'rxjs';
 
 import { environment } from '../../environments/environment';
@@ -10,40 +10,30 @@ import { AuthService } from './auth.service';
 })
 export class HubService {
   private hub: signalR.HubConnection;
-  private timer: any;
   message: BehaviorSubject<any> = new BehaviorSubject({});
 
   constructor(private auth: AuthService) {}
 
   async start() {
-    if (!this.auth.isAuthorized) {
+    if (!this.auth.isAuthorized()) {
       return;
     }
-
     this.hub = new signalR
       .HubConnectionBuilder()
       .withUrl(`${environment.hubUrl}`, { accessTokenFactory: () => this.auth.getToken() })
+      .withAutomaticReconnect()
       .build();
 
-    await this.connect();
+    await this.hub.start();
 
-    this.hub.onclose(async () => {
-      await this.connect();
-    });
+    this.hub.onclose(async () => await this.hub.start());
+
+    this.hub.onreconnecting(error => console.log('reconnecting signalR ', error));
 
     this.hub.on('notify', (msg: any) => this.message.next(msg));
   }
 
-  private async connect() {
-    try {
-      await this.hub.start();
-    } catch {
-      if (this.timer) {
-        clearInterval(this.timer);
-      }
-      this.timer = setInterval(async () => {
-        await this.hub.start();
-      }, 5000);
-    }
+  async stop() {
+    return this.hub && await this.hub.stop();
   }
 }
